Fix bug where not having api/v2 directory causes PHPStan crash.
Now the PHPStanConfig script will always generate the necessary file,
even if it doesn't have anything to put in the file. It also won't be as
alarming with its error message, given that things should still work in
this case (but would be unusual in the future-compatible case where the
system should have an api/v2 folder).