Pretty decent API V2 create-checkout endpoint definition
It does still have some caveats that are highlighted with TODO comments
in the StripeService class. Things like "maybe we should be transmitting
string for our monetary amounts because integers/numbers might not have
enough range?" or "maybe we should use create a new class with
ClassOfConstantIntegersTrait to store currency definitions?", and
others.