Bugfix in phpstan.sh script
This bug caused phpstan.sh to fail to pick up the content of
custom `phpstan-opts.txt` and `phpstan-paths.txt` if it wasn't executed
from the project root directory.

It was caused by using an undeclared path variable near the beginning of
the script. With the typo fixed, this is no longer a problem.