SemVer - Semantic Version handling module
This module contains a number of functions for working with
semantic-versions-as-strings: construction/composing,
deconstruction/decomposing/extraction, equality, comparison, and _some_
wildcard pattern matching.

It does fail a few unittests at the moment, but they are obscure enough
that it's worthwhile to just commit the code. These unittest failures
will print messages during testing and they have been marked with TODO
tags, so it should be easy enough to come back to this later.