Internally usable `StripeService::create_checkout` function.
This is the result of splitting `api_create_checkout` into two
functions:
* One, still called `api_create_checkout`, that is suitable for handling
  endpoint requests for creating Stripe checkouts.
* Another, called `create_checkout`, which can be used by other Kickback
  code (e.g. different endpoints that need Stripe integration) to send
  Stripe checkout requests.

This commit also adds a `kk_needs_session` field to the JSON request
format. This allows us to turn off the session requirements during
testing. The field is ignored if the server has a hostname
that isn't `localhost` or `127.0.0.1`, as determined by the
Stripe redirect URI in the config file (e.g. it should be impossible
for anyone to use this field when connecting to a proper server
instance on the internet: it will just be ignored and a session
required regardless).